Move cmd/unused/ui to cmd/unused/internal/ui
This directory isn't meant to be exported.
