+++ cmd/unused/internal/ui/interactive/model.go
