+++ cmd/unused/internal/ui/interactive/provider_view.go
